pyproject.toml
[tool.poetry]
name = "brendwall"
version = "0.1.0"
description = ""
authors = ["Anton <patrin_toxa@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
